+++ src/main/java/io/github/arthursilvagbs/Locacao/de/Carros/dto/veiculo/VeiculoUpdateDTO.java
@@ -1,7 +1,7 @@
 package io.github.arthursilvagbs.Locacao.de.Carros.dto.veiculo;
 
 public record VeiculoUpdateDTO(
-        Integer quilometragem,
+        Double quilometragem,
         String cor,
         String placa
 ) {
